pyproject.toml
[tool.poetry]
name = "trier-de-news-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Christopher Perrin <cperrin88@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
feedparser = "^6.0.10"
python-telegram-bot = "^13.15"
click = "^8.1.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^0.991"
pytest = "^7.2.0"
pylint = "^2.15.8"
black = "^22.12.0"
types-pytz = "^2022.6.0.1"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "feedparser.*"
ignore_missing_imports = true

[tool.poetry.scripts]
trier-bot = 'trier_de_news_telegram_bot:main'
